Move preferences to aws module
