+++ aws/src/main/kotlin/edu/wpi/axon/aws/preferences/Preferences.kt
@@ -1,17 +1,18 @@
-package edu.wpi.axon.preferences
+package edu.wpi.axon.aws.preferences
 
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.JsonConfiguration
+import software.amazon.awssdk.services.ec2.model.InstanceType
 
 /**
- * The user's preferences.
+ * The user's edu.wpi.axon.aws.preferences.
  *
  * @param defaultEC2NodeType The default EC2 node type to run training scripts in.
  */
 @Serializable
 data class Preferences(
-    val defaultEC2NodeType: String = "t2.micro"
+    val defaultEC2NodeType: InstanceType = InstanceType.T2_MICRO
 ) {
 
     fun serialize(): String = Json(
