--- preferences/preferences.gradle.kts
@@ -1,5 +1,0 @@
-description = "The preferences API interfaces."
-
-dependencies {
-    implementation(project(":logging"))
-}
