+++ settings.gradle.kts
@@ -40,7 +40,6 @@ include(":dsl-test-util")
 include(":example-models")
 include(":logging")
 include(":pattern-match")
-include(":preferences")
 include(":tasks-yolov3")
 include(":test-util")
 include(":tf-data")
