+++ ui-vaadin/src/main/kotlin/edu/wpi/axon/ui/view/preferences/PreferencesView.kt
@@ -1,25 +1,31 @@
 package edu.wpi.axon.ui.view.preferences
 
 import com.github.mvysny.karibudsl.v10.KComposite
-import com.github.mvysny.karibudsl.v10.h1
-import com.github.mvysny.karibudsl.v10.horizontalLayout
-import com.github.mvysny.karibudsl.v10.icon
-import com.github.mvysny.karibudsl.v10.text
+import com.github.mvysny.karibudsl.v10.comboBox
+import com.github.mvysny.karibudsl.v10.formLayout
 import com.github.mvysny.karibudsl.v10.verticalLayout
-import com.vaadin.flow.component.dependency.NpmPackage
-import com.vaadin.flow.component.icon.VaadinIcon
-import com.vaadin.flow.component.orderedlayout.FlexComponent
 import com.vaadin.flow.router.PageTitle
 import com.vaadin.flow.router.Route
-import com.vaadin.shrinkwrap.VaadinCoreShrinkWrap
+import edu.wpi.axon.aws.preferences.PreferencesManager
 import edu.wpi.axon.ui.MainLayout
+import org.koin.core.KoinComponent
+import org.koin.core.inject
+import software.amazon.awssdk.services.ec2.model.InstanceType
 
-@Route(value = "preferences", layout = MainLayout::class)
+@Route(value = "", layout = MainLayout::class)
 @PageTitle("Preferences")
-class PreferencesView : KComposite() {
+class PreferencesView : KComposite(), KoinComponent {
+    private val preferencesManager by inject<PreferencesManager>()
+
     private val root = ui {
         verticalLayout {
-
+            formLayout {
+                comboBox<InstanceType>("Training Instance Type") {
+                    isAllowCustomValue = true
+                    placeholder = InstanceType.T2_MICRO.toString()
+                    isRequired = true
+                }
+            }
         }
     }
 }
